requirements.txt
django
psycopg2-binary
django-ninja
django-ninja-simple-jwt
pillow